[v0.3.1] Refactor: Reorganize workspace components into feature-based structure
Reorganize the workspace components into logical feature-based directories (explore, mod-queue, mod-manager, shared) for better maintainability. Break down the large InstalledTab component into focused sub-components: InstalledModsList, InstalledUpdatesList, InstalledVersionDropdown, and InstalledModals. Extract modal logic and reusable utilities. Update all import paths accordingly. Remove unused assets and demo components.
